# internal/server/openapi.yaml
openapi: "3.1.0"
info:
  title: ob serve API
  version: 0.1.0
  description: Local HTTP API for the OpenBindings CLI (ob serve).
servers:
  - url: ${OB_SERVER_URL}
    description: Default local server (port 20290 = 0x4F42, ASCII OB)
paths:
...
  /oauth/authorize:
    get:
      operationId: oauthAuthorize
      summary: OAuth2 authorization endpoint (PKCE required)
      description: >
        Initiates the OAuth2 Authorization Code flow with PKCE. Renders a consent
        page in the browser. The user approves or denies the request, and the server
        redirects back to the client's redirect_uri with an authorization code.
      security: []
      parameters:
        - name: response_type
          in: query
          required: true
          schema:
            type: string
            enum: [code]
        - name: client_id
          in: query
          required: true
          schema:
            type: string
        - name: redirect_uri
          in: query
          required: true
          schema:
            type: string
        - name: code_challenge
          in: query
          required: true
          schema:
            type: string
          description: Base64url-encoded SHA256 hash of the code_verifier
        - name: code_challenge_method
          in: query
          required: true
          schema:
            type: string
            enum: [S256]
        - name: state
          in: query
          schema:
            type: string
      responses:
        "200":
          description: HTML consent page
          content:
            text/html:
              schema:
                type: string
    post:
      operationId: oauthApprove
      summary: OAuth2 consent form submission
      description: >
        Processes the user's approve/deny action from the consent page.
        On approval, redirects to the client's redirect_uri with an authorization code.
        On denial, redirects with an error.
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [nonce, action]
              properties:
                nonce:
                  type: string
                  description: Authorization request nonce from the consent page
                action:
                  type: string
                  enum: [approve, deny]
      responses:
        "302":
          description: Redirect to client redirect_uri with code or error
  /oauth/token:
    post:
      operationId: oauthToken
      summary: OAuth2 token endpoint (exchanges code for access token)
      security: []
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              required: [grant_type, code, code_verifier]
              properties:
                grant_type:
                  type: string
                  enum: [authorization_code]
                code:
                  type: string
                code_verifier:
                  type: string
                redirect_uri:
                  type: string
      responses:
        "200":
          description: Access token
          content:
            application/json:
              schema:
                type: object
                properties:
                  access_token:
                    type: string
                  token_type:
                    type: string
                    enum: [Bearer]
                  expires_in:
                    type: integer
  /http/request:
    post:
      operationId: request
      summary: Make an HTTP request and return the response
      description: >
        Makes an HTTP request on behalf of the caller and returns the complete
        response. Useful for clients that cannot make direct HTTP requests due
        to platform constraints (browser CSP/CORS, network restrictions, etc.).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/HttpRequestInput"
      responses:
        "200":
          description: HTTP response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HttpResponse"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Static session token (printed on startup or set via --token/OB_SERVE_TOKEN)
        or an OAuth2 access token obtained via the /oauth/authorize + /oauth/token flow.
  schemas:
...
    HttpRequestInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
          description: The URL to request
        method:
          type: string
          description: HTTP method (defaults to GET)
        headers:
          type: object
          additionalProperties:
            type: string
          description: Request headers
        body:
          type: string
          description: Request body as text
        timeoutMs:
          type: integer
          description: Request timeout in milliseconds
        maxResponseBytes:
          type: integer
          description: Maximum response body size in bytes
        followRedirects:
          type: boolean
          description: Whether to follow HTTP redirects (defaults to true)
    HttpResponse:
      type: object
      required: [status, body]
      properties:
        status:
          type: integer
          description: HTTP status code
        headers:
          type: object
          additionalProperties:
            type: string
          description: Response headers
        body:
          type: string
          description: Response body as text
        url:
          type: string
          description: Final URL after any redirects
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
          description: Human-readable error message
        detail:
          type: string
          description: Additional context about the error
        code:
          type: string
          description: Machine-readable error code
security:
  - bearerAuth: []